Allow the silencing of $stderr output.
In tests, you may not be setting the required keys etc, which means that this
gem gets super noisy in specs.

This commit allows passing in of a `silent` option which silences these messages.

There are no tests, but I couldn't find any tests that test the $stderr output
at all.